+++ R/to_factor.R
@@ -31,7 +31,7 @@ NULL
 #' @rdname to_factor
 #' @export
 
-to_factor <- function(x, ordered = TRUE, drop.levels = FALSE) {
+to_factor <- function(x, ordered = TRUE, drop_levels = FALSE) {
   if (!is.null(val_lab(x))) {
     vallab <- val_lab(x)
   } else {
@@ -44,7 +44,7 @@ to_factor <- function(x, ordered = TRUE, drop.levels = FALSE) {
     labels = names(vallab),
     ordered = ordered
   )
-  if (drop.levels) {
+  if (drop_levels) {
     res <- droplevels(res)
   }