+++ R/utf8.R
@@ -15,18 +15,18 @@ detect_invalid_utf8 <- function(data) {
   if (!inherits(data, "data.frame")) {
     stop("input a data.frame")
   }
-  X <- lapply(data, invalid_utf8_)
-  where <- lapply(X, which)
+  x <- lapply(data, invalid_utf8)
+  where <- lapply(x, which)
   col_names <- names(where)
   cols <- rep(col_names, lapply(where, length))
   rows <- unlist(where)
-  X_array <- array(unlist(X), dim = dim(data))
+  x_array <- array(unlist(x), dim = dim(data))
   # Tibble has issues, convert to data.frame
   data <- as.data.frame(data)
   data.frame(
     column = cols,
     row = rows,
-    value = unlist(data[X_array]),
+    value = unlist(data[x_array]),
     row.names = NULL
   )
 }
@@ -37,12 +37,12 @@ remove_invalid_utf8 <- function(data) {
   if (!inherits(data, "data.frame")) {
     stop("input a data.frame")
   }
-  X <- lapply(data, rm_invalid_utf8_)
-  data.frame(X)
+  x <- lapply(data, rm_invalid_utf8)
+  data.frame(x)
 }
 
 #' @keywords internal
-rm_invalid_utf8_ <- function(x) {
+rm_invalid_utf8 <- function(x) {
   x <- iconv(x, "UTF-8", "UTF-8", sub = "")
   gsub("\u00A0", "\u0020", x)
 }
@@ -50,7 +50,7 @@ rm_invalid_utf8_ <- function(x) {
 
 #' @keywords internal
 
-invalid_utf8_ <- function(x) {
+invalid_utf8 <- function(x) {
   !is.na(x) & is.na(iconv(x, "UTF-8", "UTF-8")) |
     grepl("\u00A0", x)
 }